requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
jinja2==3.1.5
httpx==0.28.1
python-dotenv==1.0.1
apscheduler==3.10.4
